refactor(setup): move YAML-specific logic to yaml.go
This commit refactors the ``internal/setup/setup.go`` file and moves the
YAML-specific logic to ``yaml.go`` within the same package.
